Added a new implementation for IR to generate SVT reliably
Earlier implementation of sytax validation tree missed a few edge cases
when more complex syntax was to be designed, to counter these edge cases
we can parse the syntax in to steps. First step will generate a basic
lazy intermediate representation of the tree. At the second step we can
parse the IR to create the tree, (Lazy evaluation will result in blocked
syntax strings to not be evaluated and be marked for later evaluation,
this evaluation of marked items will be done at tree generation). This
process may consume more resources than the previous implementation, but
will yeild much more predictable results. For now the intermediate
representation generation mechanism is ready.